Modularize the hierarchy of tests
Create a folder for integration tests.
Create a folder for models tests.
Modify rakefile to work with the new changes.
