+++ rakefile
@@ -2,8 +2,8 @@ task default: :test
 
 desc 'Run all test'
 task :test do
-  require_relative './tests/test_helper'
-  Cutest.run Dir['tests/*_test.rb']
+  require_relative './test/test_helper'
+  Cutest.run Dir['test/*/*_test.rb']
 end
 
 namespace 'seed' do
