+++ test/lib/migration.rb
