+++ test/test_helper.rb
@@ -1,7 +1,11 @@
 ENV['RACK_ENV'] = 'test'
 ENV['REDISCLOUD_URL'] = 'redis://localhost:6379/2'
 require './app'
-require './tests/migration'
+require './test/lib/migration'
+
+def body_json
+  JSON.parse last_response.body, symbolize_names: true
+end
 
 prepare do
   Ohm.flush
